thelia/doc/exemple_htaccess
git-svn-id: https://thelia.svn.sourceforge.net/svnroot/thelia@259 6d4e0a53-8826-0410-8364-8b493f4a4454

+++ thelia/admin/configuration.php
@@ -48,7 +48,7 @@
 	include_once("entete.php");
 ?>
 <?php
-	if($action == "videcache"){
+	if(isset($action) && $action == "videcache"){
 		$cache = new Cache();
 		$query = "delete from $cache->table";
 		$resul = mysql_query($query, $cache->link);
